# perf(ext/fetch): Use the WebIDL conversion to DOMString rather than USVString for Response constructor