--- a/ext/fetch/23_response.js
+++ b/ext/fetch/23_response.js
@@ -254,7 +254,7 @@
      */
     constructor(body = null, init = {}) {
       const prefix = "Failed to construct 'Response'";
-      body = webidl.converters["BodyInit?"](body, {
+      body = webidl.converters["BodyInit_DOMString?"](body, {
         prefix,
         context: "Argument 1",
       });