@lucacasonato 2021-09-23
This is not correct though. USVString normalization doesn't happen at all now.