@andreubotella 2021-09-23
> This is not correct though. USVString normalization doesn't happen at all now.

The `BodyInit` type is only used in the `Response` constructor and in `RequestInit`, which itself is only used in the `Request` constructor (and in the `fetch` function, which defers to the `Request` constructor). In both of these cases, the `BodyInit` parameter is passed to `extractBody` before doing anything else with it – and `extractBody` will UTF-8 encode strings with replacement, effectively performing the USV normalization.